🚚 Trekk `EndreDelvikår` ut i egen fil
