+++ src/frontend/Sider/Behandling/Vilkårvurdering/EndreVilkår.tsx
@@ -4,18 +4,9 @@ import styled from 'styled-components';
 
 import { TrashIcon } from '@navikt/aksel-icons';
 import { ErrorMessage, VStack } from '@navikt/ds-react';
-import { ABorderAction, AShadowXsmall } from '@navikt/ds-tokens/dist/tokens';
+import { AShadowXsmall } from '@navikt/ds-tokens/dist/tokens';
 
-import Begrunnelse from './Begrunnelse';
-import DelvilkårRadioknapper from './DelvilkårRadioknapper';
-import {
-    begrunnelseErPåkrevdOgUtfyllt,
-    hentSvaralternativ,
-    kanHaBegrunnelse,
-    kopierBegrunnelse,
-    leggTilNesteIdHvis,
-    oppdaterSvarIListe,
-} from './utils';
+import EndreDelvilkår from './EndreVilkår/EndreDelvilkår';
 import { Feilmeldinger, ingen, ingenFeil, validerVilkårsvurderinger } from './validering';
 import { useApp } from '../../../context/AppContext';
 import { useBehandling } from '../../../context/BehandlingContext';
@@ -23,7 +14,7 @@ import SmallButton from '../../../komponenter/Knapper/SmallButton';
 import { Skillelinje } from '../../../komponenter/Skillelinje';
 import { SmallWarningTag } from '../../../komponenter/Tags';
 import { FlexColumn } from '../../../komponenter/Visningskomponenter/Flex';
-import { BegrunnelseRegel, Regler, Svaralternativ } from '../../../typer/regel';
+import { Regler } from '../../../typer/regel';
 import { RessursFeilet, RessursStatus, RessursSuksess } from '../../../typer/ressurs';
 import {
     Delvilkår,
@@ -31,24 +22,10 @@ import {
     StønadsvilkårType,
     Vilkår,
     Vilkårsresultat,
-    Vurdering,
 } from '../vilkår';
 import EndrePeriodeForVilkår, {
     EndrePeriodeForVilkårForm,
 } from './EndreVilkår/EndrePeriodeForVilkår';
-
-const DelvilkårContainer = styled.div<{ $erUndervilkår: boolean }>`
-    border-left: ${({ $erUndervilkår }) =>
-        $erUndervilkår ? `5px solid ${ABorderAction}` : 'none'};
-    padding-left: ${({ $erUndervilkår }) => ($erUndervilkår ? '1rem' : '0')};
-    gap: ${({ $erUndervilkår }) => ($erUndervilkår ? `5rem` : `6rem`)};
-    display: flex;
-
-    @media (max-width: 900px) {
-        flex-direction: column;
-        gap: 1rem;
-    }
-`;
 
 const StyledForm = styled.form`
     background: white;
@@ -136,82 +113,6 @@ export const EndreVilkår: FC<EndreVilkårProps> = (props) => {
         // eslint-disable-next-line react-hooks/exhaustive-deps
     }, [detFinnesUlagredeEndringer]);
 
-    const oppdaterVilkårsvar = (index: number, nySvarArray: Vurdering[]) => {
-        settDelvilkårsett((prevSvar) => {
-            const prevDelvilkårsett = prevSvar[index];
-            return [
-                ...prevSvar.slice(0, index),
-                {
-                    ...prevDelvilkårsett,
-                    vurderinger: nySvarArray,
-                },
-                ...prevSvar.slice(index + 1),
-            ];
-        });
-    };
-
-    const oppdaterBegrunnelse = (
-        vurderinger: Vurdering[],
-        delvilkårIndex: number,
-        nyttSvar: Vurdering
-    ) => {
-        const { begrunnelse } = nyttSvar;
-        const svaralternativ: Svaralternativ | undefined = hentSvaralternativ(
-            props.regler,
-            nyttSvar
-        );
-        if (!svaralternativ) {
-            return;
-        }
-
-        const oppdaterteSvar = oppdaterSvarIListe(nyttSvar, vurderinger, true);
-
-        const oppdaterteSvarMedNesteRegel = leggTilNesteIdHvis(
-            svaralternativ.regelId,
-            oppdaterteSvar,
-            () => begrunnelseErPåkrevdOgUtfyllt(svaralternativ, begrunnelse)
-        );
-        oppdaterVilkårsvar(delvilkårIndex, oppdaterteSvarMedNesteRegel);
-    };
-
-    const oppdaterSvar = (
-        vurderinger: Vurdering[],
-        delvilkårIndex: number,
-        nyttSvar: Vurdering
-    ) => {
-        const svaralternativer: Svaralternativ | undefined = hentSvaralternativ(
-            props.regler,
-            nyttSvar
-        );
-
-        if (!svaralternativer) {
-            return;
-        }
-
-        const oppdaterteSvar = oppdaterSvarIListe(
-            nyttSvar,
-            vurderinger,
-            false,
-            kanHaBegrunnelse(svaralternativer)
-        );
-
-        const oppdaterteSvarMedNesteRegel = leggTilNesteIdHvis(
-            svaralternativer.regelId,
-            oppdaterteSvar,
-            () => svaralternativer.begrunnelseType !== BegrunnelseRegel.PÅKREVD
-        );
-
-        const oppdaterteSvarMedKopiertBegrunnelse = kopierBegrunnelse(
-            vurderinger,
-            oppdaterteSvarMedNesteRegel,
-            nyttSvar,
-            svaralternativer,
-            props.regler
-        );
-
-        oppdaterVilkårsvar(delvilkårIndex, oppdaterteSvarMedKopiertBegrunnelse);
-    };
-
     const validerOgLagreVilkårsvurderinger = async (event: React.FormEvent<HTMLFormElement>) => {
         event.preventDefault();
         const { fom, tom, utgift, erFremtidigUtgift } = endrePeriodeForVilkårForm;
@@ -245,51 +146,6 @@ export const EndreVilkår: FC<EndreVilkårProps> = (props) => {
         }
     };
 
-    const nullstillFeilmeldingForRegel = (regelId: string) => {
-        settFeilmeldinger({
-            ...feilmeldinger,
-            delvilkårsvurderinger: { ...feilmeldinger.delvilkårsvurderinger, [regelId]: undefined },
-        });
-    };
-
-    const EndreDelvilkår = delvilkårsett.map((delvikår, delvilkårIndex) => {
-        return delvikår.vurderinger.map((svar) => {
-            const gjeldendeRegel = props.regler[svar.regelId];
-            const erUndervilkår = !gjeldendeRegel.erHovedregel;
-            return (
-                <React.Fragment key={gjeldendeRegel.regelId}>
-                    {delvilkårIndex !== 0 && !erUndervilkår && <Skillelinje />}
-                    <DelvilkårContainer $erUndervilkår={erUndervilkår}>
-                        <DelvilkårRadioknapper
-                            vurdering={svar}
-                            regel={gjeldendeRegel}
-                            readOnly={!props.alleFelterKanRedigeres}
-                            settVurdering={(nyVurdering) => {
-                                settDetFinnesUlagredeEndringer(true);
-                                oppdaterSvar(delvikår.vurderinger, delvilkårIndex, nyVurdering);
-                            }}
-                            feilmelding={
-                                feilmeldinger.delvilkårsvurderinger[gjeldendeRegel.regelId]
-                            }
-                            nullstillFeilmelding={nullstillFeilmeldingForRegel}
-                        />
-                        <Begrunnelse
-                            oppdaterBegrunnelse={(begrunnelse) => {
-                                settDetFinnesUlagredeEndringer(true);
-                                oppdaterBegrunnelse(delvikår.vurderinger, delvilkårIndex, {
-                                    ...svar,
-                                    begrunnelse,
-                                });
-                            }}
-                            vurdering={svar}
-                            regel={gjeldendeRegel}
-                        />
-                    </DelvilkårContainer>
-                </React.Fragment>
-            );
-        });
-    });
-
     const slettVilkår = props.slettVilkår;
 
     return (
@@ -306,7 +162,17 @@ export const EndreVilkår: FC<EndreVilkårProps> = (props) => {
                     settFeilmeldinger={settFeilmeldinger}
                 />
                 <Skillelinje />
-                {!endrePeriodeForVilkårForm.erFremtidigUtgift && EndreDelvilkår}
+                {!endrePeriodeForVilkårForm.erFremtidigUtgift && (
+                    <EndreDelvilkår
+                        delvilkårsett={delvilkårsett}
+                        regler={props.regler}
+                        alleFelterKanRedigeres={props.alleFelterKanRedigeres}
+                        settDetFinnesUlagredeEndringer={settDetFinnesUlagredeEndringer}
+                        feilmeldinger={feilmeldinger}
+                        settFeilmeldinger={settFeilmeldinger}
+                        settDelvilkårsett={settDelvilkårsett}
+                    />
+                )}
                 <VStack gap="4">
                     <Knapper>
                         <SmallButton>Lagre</SmallButton>
